# application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于区块链的钢材供应链融资解决方案api文档",
        "title": "基于区块链的钢材供应链融资解决方案",
        "contact": {
            "name": "togettoyou",
            "email": "144240096@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/DeleteProduct": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "更改订单状态",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductByNameQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/QueryProduct": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询全部货品",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryProductByGoodType": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "按货品类型查询",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductByGoodTypeQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryProductByID": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "按货品ID查询",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductByNameQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryProductByMaterial": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "按材料查询",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductByMaterialQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryProductBySupplier": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "按供应商查询",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductBySupplierQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryProductByWorkmanship": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "按制作工艺查询",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductByWorkmanshipQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/StoreProduct": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "存储到数据库",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/UpdateProductByAmount": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新货品数量",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductByAmountUpdateRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "v1.ProductByAmountUpdateRequestBody": {
            "type": "object",
            "properties": {
                "Amount": {
                    "description": "货品数量",
                    "type": "integer"
                },
                "GoodsID": {
                    "description": "货品ID",
                    "type": "string"
                }
            }
        },
        "v1.ProductByGoodTypeQueryRequestBody": {
            "type": "object",
            "properties": {
                "GoodType": {
                    "description": "货品类型",
                    "type": "string"
                }
            }
        },
        "v1.ProductByMaterialQueryRequestBody": {
            "type": "object",
            "properties": {
                "Material": {
                    "description": "材料",
                    "type": "string"
                }
            }
        },
        "v1.ProductByNameQueryRequestBody": {
            "type": "object",
            "properties": {
                "GoodsID": {
                    "description": "货品ID",
                    "type": "string"
                }
            }
        },
        "v1.ProductBySupplierQueryRequestBody": {
            "type": "object",
            "properties": {
                "Supplier": {
                    "description": "供应商",
                    "type": "string"
                }
            }
        },
        "v1.ProductByWorkmanshipQueryRequestBody": {
            "type": "object",
            "properties": {
                "Workmanship": {
                    "description": "制作工艺",
                    "type": "string"
                }
            }
        },
        "v1.ProductRequestBody": {
            "type": "object",
            "properties": {
                "AddDate": {
                    "description": "上架时间",
                    "type": "string"
                },
                "Amount": {
                    "description": "货品数量",
                    "type": "integer"
                },
                "GoodType": {
                    "description": "货品类型",
                    "type": "integer"
                },
                "GoodsID": {
                    "description": "货品ID",
                    "type": "string"
                },
                "Material": {
                    "description": "材料",
                    "type": "string"
                },
                "Name": {
                    "description": "货品名称",
                    "type": "string"
                },
                "Price": {
                    "description": "价格",
                    "type": "integer"
                },
                "Remark": {
                    "description": "备注",
                    "type": "string"
                },
                "ShippingArea": {
                    "description": "发货地区",
                    "type": "string"
                },
                "Standards": {
                    "description": "规格",
                    "type": "string"
                },
                "Supplier": {
                    "description": "供应商",
                    "type": "string"
                },
                "SupplierID": {
                    "type": "string"
                },
                "Tel": {
                    "description": "联系方式",
                    "type": "string"
                },
                "Workmanship": {
                    "description": "制作工艺",
                    "type": "string"
                }
            }
        },
...
    }
}